+++ spring/src/main/java/de/taimos/daemon/spring/conditional/BeanAvailable.java
@@ -11,11 +11,11 @@ import java.lang.annotation.Target;
 
 import org.springframework.context.annotation.Conditional;
 
-@Target(ElementType.TYPE)
+@Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Conditional(BeanAvailableCondition.class)
 public @interface BeanAvailable {
 
-    Class value();
+	Class value();
 
 }